pyproject.toml
[project]
name = "gemini-watermark-cleaner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "iopaint>=1.6.0",
    "opencv-python>=4.11.0.86",
    "pillow>=9.5.0",
    "pyqt6>=6.10.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
